Day11 Coninuation
Tried pointers to the 2d array and it was a mistake. Got it working but it looped infinitely as far as I can tell, so maybe it didn't work.
Using numFlashes as a pointer now and instead returning the 2d vector at the end of recursion.
Now I just need to account for indices that aren't bordered on all sides.